Merge 1.1.0

+++ build.sh
@@ -22,7 +22,7 @@ success()
 error()
 {
 	echo
-	echo "${red}==>${bold} Error: $1${reset}"
+	echo "${red}==>${bold} $1${reset}"
 }
 
 heading "Installing arkstats-reporter-1.0.0..."
@@ -67,49 +67,61 @@ sudo chmod 755 /etc/cron.hourly/ntpdate
 success "System time synchronised and updates scheduled."
 echo
 sleep 1
-
-heading "Pulling latest configuration file..."
-echo
-wget https://raw.githubusercontent.com/dafty-1/arkstats-reporter/master/app.json -O app.json
-
-sleep 3
 clear
 
-# Config
+if ! grep -q "\"RPC_HOST\"        : \"\"," app.json; then
+    error "ArkStats has already been configured (username, secret key, etc). Would you like to reconfigure it? [y/N]."
+    echo "If you are updating ArkStats, this is usually not required."
+    read -e -r -p ": " RECONFIGURE
+    if [[ $RECONFIGURE == "y" || $RECONFIGURE == "Y" || $RECONFIGURE == "yes" || $RECONFIGURE == "YES" || $RECONFIGURE == "Yes" ]]
+    then
+        heading "Pulling latest configuration file..."
+        echo
+        wget https://raw.githubusercontent.com/dafty-1/arkstats-reporter/master/app.json -O app.json
+    fi
+fi
 
-heading "Enter the IP address of your Ark Node installation, without quotes, followed by ENTER."
-echo "This is usually ${bold}localhost${reset}"
+if grep -q "\"RPC_HOST\"        : \"\"," app.json; then
+
+    #Config
+ 
+    heading "Enter the IP address of your Ark Node installation, without quotes, followed by ENTER."
+
+    echo "This is usually ${bold}localhost${reset}"
     read -e -r -p ": " RPC_HOST
-    sed -i "/.*RPC_HOST.*/c\ \ \ \ \ \ \"RPC_HOST\"\ \ \ \ \ \ \ \ :\ \"$RPC_HOST\"," app.json
-
-heading "Enter the port of your Ark Node installation, without quotes, followed by ENTER."
-echo "This is usually ${bold}4001${reset}"
+        sed -i "/.*RPC_HOST.*/c\ \ \ \ \ \ \"RPC_HOST\"\ \ \ \ \ \ \ \ :\ \"$RPC_HOST\"," app.json
+    
+    heading "Enter the port of your Ark Node installation, without quotes, followed by ENTER."
+    echo "This is usually ${bold}4001${reset}"
 
     read -e -r -p ": " RPC_PORT
-    sed -i "/.*RPC_PORT.*/c\ \ \ \ \ \ \"RPC_PORT\"\ \ \ \ \ \ \ \ :\ $RPC_PORT," app.json
-    sed -i "/.*LISTENING_PORT.*/c\ \ \ \ \ \ \"LISTENING_PORT\"\ \ :\ $RPC_PORT," app.json
+        sed -i "/.*RPC_PORT.*/c\ \ \ \ \ \ \"RPC_PORT\"\ \ \ \ \ \ \ \ :\ $RPC_PORT," app.json
+        sed -i "/.*LISTENING_PORT.*/c\ \ \ \ \ \ \"LISTENING_PORT\"\ \ :\ $RPC_PORT," app.json
+    
+    heading "Enter a username to identify your node, without quotes, followed by ENTER."
+    echo "This can be a delegate name, Ark address or Slack username"
+    
+        read -e -r -p ": " INSTANCE_NAME
+        sed -i "/.*INSTANCE_NAME.*/c\ \ \ \ \ \ \"INSTANCE_NAME\"\ \ \ :\ \"$INSTANCE_NAME\"," app.json
+    
+    heading "Enter an email address or website for your node, without quotes, followed by ENTER."
+    echo "This is not required but can be helpful for other users"
+    
+        read -e -r -p ": " CONTACT_DETAILS
+        sed -i "/.*CONTACT_DETAILS.*/c\ \ \ \ \ \ \"CONTACT_DETAILS\"\ :\ \"$CONTACT_DETAILS\"," app.json
 
-heading "Enter a username to identify your node, without quotes, followed by ENTER."
-echo "This can be a delegate name, Ark address or Slack username"
+    heading "Enter the secret token used to authenticate with the server, without quotes, followed by ENTER."
+    echo "Start a direct message with dafty on the ArkEcosystem Slack to get the secret token"
 
-    read -e -r -p ": " INSTANCE_NAME
-    sed -i "/.*INSTANCE_NAME.*/c\ \ \ \ \ \ \"INSTANCE_NAME\"\ \ \ :\ \"$INSTANCE_NAME\"," app.json
+        read -e -r -p ": " WS_SECRET
+        sed -i "/.*WS_SECRET.*/c\ \ \ \ \ \ \ \"WS_SECRET\"\ \ \ \ \ \ :\ \"$WS_SECRET\"," app.json
+fi
 
-heading "Enter an email address or website for your node, without quotes, followed by ENTER."
-echo "This is not required but can be helpful for other users"
-
-    read -e -r -p ": " CONTACT_DETAILS
-    sed -i "/.*CONTACT_DETAILS.*/c\ \ \ \ \ \ \"CONTACT_DETAILS\"\ :\ \"$CONTACT_DETAILS\"," app.json
-
-heading "Enter the secret token used to authenticate with the server, without quotes, followed by ENTER."
-echo "Start a direct message with dafty on the ArkEcosystem Slack to get the secret token"
-
-    read -e -r -p ": " WS_SECRET
-    sed -i "/.*WS_SECRET.*/c\ \ \ \ \ \ \ \"WS_SECRET\"\ \ \ \ \ \ :\ \"$WS_SECRET\"," app.json
-
-success "Configuration complete! Starting ArkStats for the first time..."
+success "Configuration complete! Starting ArkStats..."
 sleep 3
 
+NODE_VER=`node -v`
+sudo env PATH=$PATH:/home/ark/.nvm/versions/node/$NODE_VER/bin /usr/local/lib/node_modules/pm2/bin/pm2 unstartup systemd -u $USER --hp /home/$USER
 pm2 stop all
 pm2 delete all
 pm2 flush
@@ -120,12 +132,13 @@ pm2 start app.json
 sleep 3
 clear
 
-#success "Installing ArkStats on boot..."
-#sleep 3
+success "Installing ArkStats on boot..."
+sleep 3
 
-#sudo pm2 startup
-#pm2 save
-#sleep 6
+sudo env PATH=$PATH:/home/$USER/.nvm/versions/node/$NODE_VER/bin /usr/local/lib/node_modules/pm2/bin/pm2 startup systemd -u $USER --hp /home/$USER
+pm2 save
+
+sleep 5
 clear
 
 pm2 status
@@ -133,13 +146,12 @@ echo
 echo
 echo
 success "Installation successful!"
-success "If you have made a mistake with any of the details above, run this script again."
+success "If you have made a mistake with any of the details, run this script again."
 echo
-echo "${cyan}Please start ArkStats by typing ${bold}pm2 start app.json${reset}"
-echo "${cyan}Check the status of ArkStats anytime by typing ${bold}pm2 log${reset}"
+echo "${cyan}${bold}ArkStats is running and will be restarted automatically on boot.${reset}"
+echo "${cyan}Check the status by typing ${bold}pm2 status${reset}${cyan} or view the logs by typing ${bold}pm2 log${reset}."
 echo
-echo
-heading "Thanks for installing ArkStats. For bugs, questions or comments, please message dafty on Slack!"
+heading "For bugs, questions or comments, please message dafty on Slack!"
 echo "This script will now exit..."
 
-sleep 10
+sleep 3
